requirements.txt
asgiref==3.9.1
camelot-py==1.0.0
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click
cryptography==45.0.5
Django
django-cors-headers
django-filter
djangorestframework
et_xmlfile==2.0.0
ghostscript
numpy
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
pdfminer.six==20250506
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.22
pypdfium2==4.30.0
python-barcode==0.15.1
python-dateutil==2.9.0.post0
pytz==2025.2
qrcode==8.2
six==1.17.0
sqlparse==0.5.3
tabulate==0.9.0
tzdata==2025.2
celery
django-celery-beat
